+++ src/pretalx/schedule/models/schedule.py
@@ -54,7 +54,7 @@ class Schedule(PretalxModel):
 
     class urls(EventUrls):
         public = "{self.event.urls.schedule}v/{self.url_version}/"
-        widget_data = "{public}widget/v2.json"
+        widget_data = "{public}widget.json"
         nojs = "{public}nojs"
 
     @transaction.atomic