+++ src/tests/common/test_common_css.py
@@ -57,15 +57,12 @@ def test_regenerate_css(event):
     from pretalx.common.tasks import regenerate_css
 
     event.primary_color = "#00ff00"
-    event.settings.widget_css_checksum = "placeholder"
     event.save()
     regenerate_css(event.pk)
     event = Event.objects.get(pk=event.pk)
     for local_app in ["agenda", "cfp"]:
         assert event.settings.get(f"{local_app}_css_file")
         assert event.settings.get(f"{local_app}_css_checksum")
-    assert event.settings.widget_css_checksum != "placeholder"
-    assert event.settings.widget_css
 
 
 @pytest.mark.django_db