+++ src/tests/orga/views/test_orga_views_event.py
@@ -194,7 +194,7 @@ def test_add_custom_css_as_administrator(event, administrator_client, path):
 def test_add_logo(event, orga_client):
     assert not event.logo
     response = orga_client.get(event.urls.base, follow=True)
-    assert '<img "src="/media' not in response.content.decode()
+    assert '<img loading="lazy" "src="/media' not in response.content.decode()
     with open("../assets/icon.png", "rb") as logo:
         response = orga_client.post(
             event.orga_urls.edit_settings,
@@ -221,7 +221,7 @@ def test_add_logo(event, orga_client):
     assert response.status_code == 200
     assert event.logo
     response = orga_client.get(event.urls.base, follow=True)
-    assert '<img src="/media' in response.content.decode(), response.content.decode()
+    assert '<img loading="lazy" src="/media' in response.content.decode(), response.content.decode()
 
 
 @pytest.mark.django_db